Close StandardSetting memory leaks using destructors (#1141)
* Close memory leak in StandardSetting

The destructor for StandardSetting released memory
for m_children and m_targets members, but failed
to release the memory for m_storage. Each class
using StandardSetting as a base uses 'new' to
create storage, and the constructor assigns
that allocated memory to m_storage.

This modification adds code to the destructor
to delete m_storage. This closes the memory leak.

Resolves #1120

* Safer close of memory leak in StandardSetting

Boolean 'm_newdStorage' has been added to the
StandardSetting class. This will be set to true
when 'new' was used to allocate the memory which
is being put into Storage. In the destructor for
StandardSetting, a test has been added so that
such memory will only be deleted if m_newdStorage
is true. This allows us to free allocated memory
associated with StandardSetting, without trying
to free memory which wasn't allocated from the heap.

All of the classes derived from StandardSetting which
use 'new' to allocate the stored contents have been
updated to set 'm_newdStorage' to true. This will
ensure that upon destruction, that memory will be
freed.

Resoves: #1120

* Close memory leak with smart pointers

All of the classes derived from StandardSetting which
used to use 'new' to allocate memory have been
switched to use std::make_shared<> to create an
instance with a smart pointer. When the last reference
to such a smart pointer is destructed, the memory
is freed. This eliminates StandardSetting related
memory leaks.

Resolves: #1120

* Close memory leaks by using destructors

All of the classes derived from StandardSetting which
use 'new' to allocate memory have been updated to add
destructors which call "delete GetStorage();" This
frees up memory allocated from the heap when an
instance of that class falls out of scope or is
deleted. This closes the StandardSetting memory
leaks.

Resolves: #1120