Fix for services that download a file
The services that download a file (Content and Guide), fail if they are
the first service called after a restart of mythbackend. The QFileInfo
class is not registered with Qt Metadata, causing the failure. Add a
registration of QFileInfo in initialization of Content and Guide
services to resolve this.

Refs #1144

(cherry picked from commit 20888721af431deea96a41fe14660ef0161fba88)