Remove unnecessary patterns from mythtv/external/.gitignore
config.h and version.h are used filenames in subdirectories
of mythtv/external.

config.* is only ignored in the mythtv/ directory, not is
subdirectories, by mythtv/.gitignore by using `/config.*`,
so `!config.*` is unnecessary.

version.h is also excluded in libmythbase/.gitignore, which
does not apply to files in mythtv/external, but that filename
is used in FFmpeg.

config.mak is excluded by FFmpeg/ffbuild/.gitignore.

config.h is excluded by FFmpeg/.gitignore, but that filename
is used in libexiv2.  I think it is clearer if the .gitignore
rule is where the file is generated.

avconfig.h is excluded by FFmpeg/libavutil/.gitignore.

FFmpeg/.gitignore ignores *_g.

Nothing creates FFmpeg.old.

All other rules are already in FFmpeg/.gitignore.