Add option to default to Last Captions State
This solution adds a LastCaptions entry into the Settings
database. This boolean gets updated each time a user toggles
captions on or off, and indicates whether captions were
enabled or disabled the last time it changed. The
m_captionsEnabledbyDefault variable is switched from a
boolean to an unsigned int so that it can hold 0, 1, or 2.
It's set to 2 when the Last Captions State configuration
is selected. When starting up viewing of a video recording
or Live TV, captions are either disabled (0), enabled (1),
or set to match the last caption state (2).

Resolves: #1154