libudfread.pro: Remove win32-msvc* force include of compat.h
/TP specifies that all files are C++, which is incorrect
since libudfread is entirely C code.

The only file that uses symbols from <fcntl.h> already includes
it, so force including it to every file makes no sense.  If
HAVE_FCNTL_H is not being set properly, that is a different issue.

The only used symbol from compat.h was ssize_t.  Instead of using
C++ using, use macro definitions, so it is not necessary to modify
the code with a typedef.