clazy: Fix readlock-detaching warnings by using container ".value()".
Qt container operator[] functions (on non-const containers) will
always create a new entry in the container if it doesn't already
exist.  This makes sense when the container is on the left side of an
assignment (its being written), but is counter-intuitive when the
container is on the right side of an assignment (its being read).  Use
the value() function instead of operator[] as it will never create a
new entry in a container.