clazy: Fix readlock-detaching warnings by using range-for const loops.
Converting these for loops to operate on a const container means that
the begin/end functions can never modify the container.  Converting
them to range-for loops simplifies the code a little bit.  Also mark a
couple of SocketHandler functions as "const" so they can be called
when working with a const object.