Fixed speaker output channel order for non-standard layouts
For some cases in macOS the channel order may not be a standard kAudioChannelLayoutTag layout order.
This can happen e.g. in 5.1 audio where HDMI can swap the C and LFE channels due to ambiguous standards.
This change pulls the channel order from the device and compares it against the standard layout for the
relevant speaker configuration: 2 (stereo), 6 (5.1) or 8 (7.1) and swaps the channel map if they are in a different
order.

Tested on MacStudio running Sequoia 15.5 connected to 5.1 A/V receiver via HDMI which had C and LFE channels
swapped.