Always update video bottom info panel text
When browsing the list of videos the bottom info panel is updated
with the details of the selected video.
In commit 53d32e054c9f8dbbc61969cd589b1b9500100aef this update was
made conditional on the IsVisible() status of the ButtonListItem.
But, after a DOWN on the bottom item of the list the next item is
still invisible when the call to update the bottom info is made.
The added check on IsVisible prevents the bottom info being updated.
This commit removes the check on IsVisible and thus restores the
behavior from before commit 53d32e054c9f8dbbc61969cd589b1b9500100aef.

Refs #1172