Fix datetime.replace(tzinfo=None) with Python 3.13
Python 3.13 fixed [gh-89039]. Now `datetime.__new__()` is also
used for `datetime.replace()`, which is called with `tzinfo=None` to
convert a time-zone aware instance to a _naive_ instance. But
`__new__()` replaces `tzinfo=None` with `loczlTZ()`, which just re-adds
the offset back.
This breaks calling `database.update()` on all database records, which
use timestamps as a datetime with the wrong timezone offset is used.

Add a new singleton LTZ and use that instead of `tzinfo=None` to keep
the old behavior to get the local timezone by default.

Fixes: #1179
[gh-89039]: https://github.com/python/cpython/issues/89039

Many thanks to github user @pmhahn providing the patch.