Qt error for MythCaptionsState
On playback of some files the following error message has been observed:
2025-10-30 12:00:13.641144 I [1196066/1196342] Decoder mythcommandlineparser.cpp:2938:operator()
    Qt: QObject::connect: Cannot queue arguments of type 'MythCaptionsState&'
    (Make sure 'MythCaptionsState&' is registered using qRegisterMetaType().)
This is now fixed in the signal declaration of CaptionsStateChanged and this
declaration does now also match the implementation in tvplaybackstate.
No differences in the behaviour have been observed other than that
the error message is not given anymore.