Seek table rebuild for videos with mythcommflag
The markup data for videos in database table filemarkup
can be created with "mythcommflag --rebuild".
This data is needed for cutlist editing and the presence of this
data may possibly also speed up navigating in playback of a video file.

The rebuild creates a record type 33, MARK_DURATION_MS, for each GOP,
each GOP identified by the mark value. The total duration of the file
is also stored as record type 33 but with mark value 0.

The problem is that the code that writes the total duration deletes many
of the type 33 records in the table. This causes problems navigating in
playback of a video file. It also causes cutlist editing to fail.

The solution as implemented does replace only the record type 33 for
mark 0 and it leaves all other records with type 33 intact.