Console output mythcommflag
Fix implementation of the --noprogress and --quiet command line
options so that it matches the documentation.
Option --noprogress controls the progress reporting which is
written directly to stdout.
Option --quiet controls the MythTV logging subsystem.
When both --noprogress and --quiet are given there is nothing
at all written to the console.