Cutlist editor audio level graph
This graph visualises the audio level +/- 0.5 frame around
the currently paused frame, as described in commit #d40411e.

This patch does fix the following issues:
- Buffer pointer computation depends on sample value size.
The audio input can have sample values of 8, 16 or 32 bits.
The samples are converted to 16 bits and then stored internally.
Computations on the input buffer need to be done with the specified
input sample size, but computations on the internal buffer must
be done with the normalized sample value size of 16 bits
instead of with the input sample value size.
- Wrong graph dispayed sometimes.
Do not create a graph if there is no audio data available
for the requested timecode.
Erase a possibly present graph if there is no graph to display.
- Graph often not displayed.
The audio data is received in one thread but the graph created
from the audio data is computed in another thread. Depending on
the thread dynamics, graph creation can fail because the audio
data has not yet arrived.
This is solved by waiting a limited amount of time until
the audio data is available before the graph is created.
- Memory leak
The memory leak was caused by the creation of a new MythImage
for each new graph that was created. This is solved by using
a single MythImage for all graphs in single cutlist edit session
and by decreasing the reference count of the MythImage when
the cutlist session is terminated.