cmake: Add support for building qt 6.10 for android/mingw.
Qt6 requires that the host and target versions be the same, so
building MythTV for android/windows on a Qt6.10 system means that you
have to compile a Qt6.10 target.

Note: Qt 6.10 has added a call to pthread_getname_np() which is not
present in the Android libraries until API version 26.  Apparently the
Qt documentation says that API26 was required for Qt 6.5, but this is
the first time there has been a compilation failure.  Fail early.