cmake/win: Add argument cast to libass sources for windows builds.
Patch the libass sources to add an explicit cast for the second
argument to iconv().  This allows setting "const" properly for the
argument, as the newer Windows cross compiler complains about passing
a "char **" value to a "const char**" argument.  This shouldn't affect
android builds as the undefined WINICONV_CONST should be treated as an
empty string.