Fix Qt6 compilation errors in qstring.h.
The complicated code in Qt6 to determine the type of an argument being
passed to QString::arg seems to have run afoul of changes to
libstdc++; changes similar to those the MythTV code fixed in
7c781b3860.  While trying to determine the type of a variable
(specifically an instance of a class), Qt6 checks to see if it can be
converted to an integer with an "operator int()" function and checks
to see if it can be converted to a string with a "toString()"
function.  On recently updated Fedora 43 and Rawhide systems, this
code is now throwing four fatal errors each time it tries to determine
if it can coerce a variable to a known type.  Its unclear if this is
because of an upgrade to Qt 6.10 or to libstdc++ 15.2.1.

Adding explicit calls to a "toInt()" function or explicitly casting
the variable to "int" isn't an elegant solution, but it solves the
compilation failures without having to wait for Qt to resolve the
problem.