Check status of programinfocache before updating
Depending on the dynamic behaviour of mythbackend and mythfrontend,
ProgramInfoCache::UpdateFileSize can be called with a recordingID
that refers to a cache entry that is being deleted.
This has been uncovered by a valgrind run with Qt 6.11 (dev) in
a debug build.
Note that there is already a check if the recordingID does point
to an existing cache entry; it only did not check if the entry
was still valid.